changes to get the 500 INR to USD

+++ src/main/java/com/example/CurrencyConverter/service/CurrencyConverter.java
@@ -2,5 +2,5 @@ package com.example.CurrencyConverter.service;
 
 public interface CurrencyConverter {
 
-    public Object convertCurrency(String fromCurrency, String toCurrency, Long units);
+    public Object convertCurrency(String fromCurrency, String toCurrency, Double units);
 }
